move unimod adding up
