+++ alphabase/psm_reader/maxquant_reader.py
@@ -8,7 +8,6 @@ import numba
 import numpy as np
 import pandas as pd
 
-from alphabase.constants.modification import MOD_DF
 from alphabase.psm_reader.keys import PsmDfCols
 from alphabase.psm_reader.psm_reader import (
     PSMReaderBase,
@@ -18,16 +17,6 @@ from alphabase.psm_reader.psm_reader import (
 
 # make sure all warnings are shown
 warnings.filterwarnings("always")
-
-mod_to_unimod_dict = {}
-for mod_name, unimod_id_ in MOD_DF[["mod_name", "unimod_id"]].to_numpy():
-    unimod_id = int(unimod_id_)
-    if unimod_id in (-1, "-1"):
-        continue
-    if mod_name[-2] == "@":
-        mod_to_unimod_dict[mod_name] = f"{mod_name[-1]}(UniMod:{unimod_id})"
-    else:
-        mod_to_unimod_dict[mod_name] = f"_(UniMod:{unimod_id})"
 
 
 @numba.njit
@@ -138,6 +127,7 @@ class MaxQuantReader(PSMReaderBase):
     """Reader for MaxQuant data."""
 
     _reader_type = "maxquant"
+    _add_unimod_to_mod_mapping = True
 
     def __init__(  # noqa: PLR0913 many arguments in function definition
         self,
@@ -217,49 +207,6 @@ class MaxQuantReader(PSMReaderBase):
             # otherwise maxquant reader will modify the dict inplace
             psm_reader_yaml["maxquant"]["modification_mapping"]
         )
-
-    def set_modification_mapping(
-        self, modification_mapping: Optional[dict] = None
-    ) -> None:
-        """Set modification mapping."""
-        super().set_modification_mapping(modification_mapping)
-        self._add_all_unimod()
-        self._extend_mod_brackets()
-        self.rev_mod_mapping = self._get_reversed_mod_mapping()
-
-    def _add_all_unimod(self) -> None:
-        for mod_name, unimod in mod_to_unimod_dict.items():
-            if mod_name in self.modification_mapping:
-                self.modification_mapping[mod_name].append(unimod)
-            else:
-                self.modification_mapping[mod_name] = [unimod]
-
-    def _extend_mod_brackets(self) -> None:
-        """Update modification_mapping to include different bracket types."""
-        for key, mod_list in list(self.modification_mapping.items()):
-            mod_set = set(mod_list)
-            # extend bracket types of modifications
-            # K(Acetyl) -> K[Acetyl]
-            # (Phospho) -> _(Phospho)
-            # _[Phospho] -> _(Phospho)
-            for mod in mod_list:
-                if mod[1] == "(":
-                    mod_set.add(f"{mod[0]}[{mod[2:-1]}]")
-                elif mod[1] == "[":
-                    mod_set.add(f"{mod[0]}({mod[2:-1]})")
-
-                if mod.startswith("_"):
-                    mod_set.add(f"{mod[1:]}")
-                elif mod.startswith("("):
-                    mod_set.add(f"_{mod}")
-                    mod_set.add(f"[{mod[1:-1]}]")
-                    mod_set.add(f"_[{mod[1:-1]}]")
-                elif mod.startswith("["):
-                    mod_set.add(f"_{mod}")
-                    mod_set.add(f"({mod[1:-1]})")
-                    mod_set.add(f"_({mod[1:-1]})")
-
-            self.modification_mapping[key] = list(mod_set)
 
     def _translate_decoy(self) -> None:
         if PsmDfCols.DECOY in self._psm_df.columns:
